requirements.txt
snowflake-snowpark-python>=1.17.0
snowflake-ml-python>=1.7.0
snowflake-connector-python>=3.6.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyYAML>=6.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
